Use /run/apps.json and load it request-time (TYO-77)
When the file changes on the disk, the UI will show the latest state.
The filename has been changed from /opt/titani/share to /run because
the former is a static one and the latter is the one that also contains
the changes from the developer, using the /datafs/apps.json.d/ directory.