dapp_json_merge.py: eliminate mass dapp restarts
When dapp-systemd-bridge is restarted to generate the new configs for
dapps, it doesn't invoke all dapps to be restarted because that would
take around 2 minutes minimum. Instead, the hard dependency has been
removed from dapp@ to dapp-systemd-bridge by changing Requires= to
Wants=, and the dapps are restarted manually.

The way it is done is to constantly watch for any changes in the json
files. When it is changed and the new, merged json is in place under
/run/apps.json, restart systemd-dapp-bridge, which is very quick now as
it only restarts the fuse mount and makes systemd to reload its configs.

When finding out what dapp to restart / stop / start, the following 4
cases are separated:
* new app in app.conf.d/. In this case, the new (by default, enabled)
service must be started.
* removed app from app.conf.d/. In this case, the already disabled
service must be stoped.
* changing an existing app under app.conf.d/. To find out which dapps
were changed, we examine the hash of the related json part. When that
changes, we restart the dapp.
* if a dapp is in failed state for some reason, but is enabled, it
should only be started again if it has a chance to run this time, which
is only when its json part was changed.

After these, nginx is reloaded to reflect new configuration.

All of this restart logic can be debugged with the following command:
   `journalctl -f -u dapp-json-merge.service`

Hopefully this gives the developer some blissful times :)