dapp@.service: do not restart dapps automatically
The decision is based on the will to not eat up users resources if a
dapp keeps failing for some reason.

Limiting the restarts to 5 lead to bad user experience on the UI. The UI
would show the dapp to flip-flop between active and inactive.

Instead of automatic restart we will go for manual restarts when a dapp
fails to start.