update requirements.txt for development needs
It's best to use the latest packages locally and use fixed versions
in production / testing environment. With this strategy tests and
released product are kept stable, while we can investigate compatibility
with newer versions during development. When an incompatible change
appears, we add that to the version the most implicity way (e.g. Django<2)

Production / testing versions are determined by the OS python package
versions.