ux-titania.bb: explicitly mention all source files
The previous hack, using tar was not cache-aware. Bitbake thought
that ux.tar is a file in this repository, and when it will change,
it will rebuild. Except, ux.tar was left on the disk, and never
changed. Or if it was deleted, a warning message was produced
that the system cannot find ux.tar, and no build was triggered.

Now we are back to the original bitbake filosophy, where each
input files must be enumerated under SRC_URI, and a rebuild
will occur if and only if one of these files change. The only
backdraw is the need to maintain this file list.