Increase datafs size to 2 gigabytes
At boot time preinstalled docker images are loaded to docker
in parallel. This means there is a point in time when all images
are on datafs twice. (one in tar format, and one in /var/lib/docker)

There is room for optimization here, to give nginx a priority
over the other preinstalled images.