swupdate: fix fstab on new root during swupdate
bitbake creates a blank fstab in rpi-titania-image-raspberrypi3.ext4
image file, and writes partition definitions to the end during creating
the .wic file. The very same image file gets copied into the .swu,
without the extra lines added to fstab, producing a dysfunct image.

This script added to .swu does update the blank fstab during upgrade
from the currently active fstab, which is known to be working.

The optimal solution would be to have a working fstab in the .swu image,
but that would involve too much hacking on the bitbake side, according
to my limited knowledge of bitbake.