generate-issue.sh: Do not flood console with /etc/issue
The networkmanager hook `generate-issue.sh` is called when
the network interfaces changed, this is how /etc/issue gets
the current IP address. Somehow it is called around 4 times
during boot, which makes /etc/issue flood the console.

This change makes generate-issue.sh able to detect if /etc/issue
changed, and if no, do not send agetty HUP signal to re-display
/etc/issue.